requirements.txt
# API Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# OpenAI - pinned to avoid proxies conflict with chromadb
openai==1.14.3
httpx==0.27.0

# Vector Database
chromadb==0.4.24

# Git integration
GitPython==3.1.43

# Environment variables
python-dotenv==1.0.1

# Data handling
pydantic==2.7.1
pydantic-settings==2.2.1

# Utilities
tiktoken==0.7.0
tenacity==8.3.0
loguru==0.7.2

# UI
streamlit==1.35.0
requests==2.32.2